test: catch errors on all pages
