+++ test/e2e/smoke.test.js
@@ -4,7 +4,7 @@
 /* eslint-env mocha */
 
 import ms from 'milliseconds'
-import {createPage, createNewPad, waitForPeerId, cleanup} from './utils'
+import {createPage, createNewPad, catchPageErrors, waitForPeerId, cleanup} from './utils'
 
 afterEach(() => {
   cleanup()
@@ -13,19 +13,6 @@ afterEach(() => {
 it('creates a pad', async () => {
   const page = await createPage()
 
-  // Catch any errors
-  const errors = []
-  page.on('error', err => errors.push(err.message)) // Emitted when the page crashes.
-  page.on('pageerror', err => errors.push(err.message)) // Emitted when an uncaught exception happens within the page.
-  page.on('requestfailed', req => {
-    errors.push(`${req.failure().errorText} ${req.url()}`)
-  })
-  page.on('response', res => {
-    if (res.status() >= 400) {
-      errors.push(`${res.status()} ${res.url()}`)
-    }
-  })
-
   // Create a pad and wait for IPFS to init.
   await createNewPad(page)
   const peerId = await waitForPeerId(page)
@@ -33,8 +20,7 @@ it('creates a pad', async () => {
 
   expect(peerId).toBeTruthy()
   expect(peerId.length).toBeGreaterThan(10)
-  errors.forEach(err => console.log(err))
-  expect(errors.length).toBe(0)
+  page.expectNoError()
 }, ms.minutes(3))
 
 it('synchronises two pads via IPFS', async () => {
@@ -60,6 +46,9 @@ it('synchronises two pads via IPFS', async () => {
   // Wait for bob's doc title to match alf's
   const bobTitleRef = await bob.$(docTitleSelector)
   const valueMatches = (input, expected) => input.value === expected
-  // waitFor fn is executed in the browser context, so context has to passed as args.
+  // waitFor fn is executed in the browser context, so context has to be passed as arg.
   await bob.waitFor(valueMatches, {/* opts */}, bobTitleRef, alfTitle)
+
+  bob.expectNoError()
+  alf.expectNoError()
 }, ms.minutes(6)) // spinning up 2 browsers and syncing pads can take a while.
